rework theme generation
