+++ src/index.scss
@@ -10,6 +10,8 @@
 // Imports all Primer files in their intended order for easy mass-inclusion.
 // Should you need specific files, you can easily use separate `@import`s.
 
+@import "./theme/index.scss";
+
 // Global requirements
 @import "./core/index.scss";
 @import "./product/index.scss";
