+++ src/theme/light.scss
@@ -1,0 +1,145 @@
+@import "../support/index.scss";
+
+:root {
+  /* variables defined in the theme */
+  --border-white: #fff;
+  --border-gray: #e1e4e8;
+  --border-gray-dark: #d1d5da;
+  --border-gray-darker: #444d56;
+  --border-blue: #0366d6;
+  --border-blue-light: #c8e1ff;
+  --border-green: #34d058;
+  --border-purple: #6f42c1;
+  --border-red: #d73a49;
+  --bg-white: #fff;
+  --bg-black: #1b1f23;
+  --bg-blue-light: #f1f8ff;
+  --bg-blue: #0366d6;
+  --bg-gray-dark: #24292e;
+  --bg-gray-light: #fafbfc;
+  --bg-gray: #f6f8fa;
+  --bg-green: #28a745;
+  --bg-green-light: #dcffe4;
+  --bg-orange: #d15704;
+  --bg-purple: #6f42c1;
+  --bg-purple-light: #f5f0ff;
+  --bg-pink: undefined;
+  --bg-red: #d73a49;
+  --bg-red-light: #ffdce0;
+  --bg-yellow: #ffd33d;
+  --bg-yellow-light: #fff5b1;
+  --bg-yellow-dark: #dbab09;
+  --text-black: #1b1f23;
+  --text-white: #fff;
+  --text-blue: #0366d6;
+  --text-gray-dark: #24292e;
+  --text-gray-light: #6a737d;
+  --text-gray: #586069;
+  --text-green: #28a745;
+  --text-orange: #a04100;
+  --text-orange-light: #e36209;
+  --text-purple: #6f42c1;
+  --text-pink: undefined;
+  --text-red: #cb2431;
+  --text-yellow: #b08800;
+
+  /* promotions: SCSS -> custom properties */
+  --border-black-fade: #{$border-black-fade};
+  --border-white-fade: #{$border-white-fade};
+  --border-gray-light: #{$border-gray-light};
+  --border-green-light: #{$border-green-light};
+  --border-red-light: #{$border-red-light};
+  --border-yellow: #{$border-yellow};
+  --border-color-button: #{$border-color-button};
+  --bg-black-fade: #{$bg-black-fade};
+  --border-white: #{$border-white};
+  --border-gray: #{$border-gray};
+  --border-gray-dark: #{$border-gray-dark};
+  --border-gray-darker: #{$border-gray-darker};
+  --border-blue: #{$border-blue};
+  --border-blue-light: #{$border-blue-light};
+  --border-green: #{$border-green};
+  --border-purple: #{$border-purple};
+  --border-red: #{$border-red};
+  --bg-white: #{$bg-white};
+  --bg-black: #{$bg-black};
+  --bg-blue-light: #{$bg-blue-light};
+  --bg-blue: #{$bg-blue};
+  --bg-gray-dark: #{$bg-gray-dark};
+  --bg-gray-light: #{$bg-gray-light};
+  --bg-gray: #{$bg-gray};
+  --bg-green: #{$bg-green};
+  --bg-green-light: #{$bg-green-light};
+  --bg-orange: #{$bg-orange};
+  --bg-purple: #{$bg-purple};
+  --bg-purple-light: #{$bg-purple-light};
+  --bg-pink: #{$bg-pink};
+  --bg-red: #{$bg-red};
+  --bg-red-light: #{$bg-red-light};
+  --bg-yellow: #{$bg-yellow};
+  --bg-yellow-light: #{$bg-yellow-light};
+  --bg-yellow-dark: #{$bg-yellow-dark};
+  --text-black: #{$text-black};
+  --text-white: #{$text-white};
+  --text-blue: #{$text-blue};
+  --text-gray-dark: #{$text-gray-dark};
+  --text-gray-light: #{$text-gray-light};
+  --text-gray: #{$text-gray};
+  --text-green: #{$text-green};
+  --text-orange: #{$text-orange};
+  --text-orange-light: #{$text-orange-light};
+  --text-purple: #{$text-purple};
+  --text-pink: #{$text-pink};
+  --text-red: #{$text-red};
+  --text-yellow: #{$text-yellow};
+
+  /* override global variables to use var() */
+  $border-black-fade: var(--border-black-fade) !global;
+  $border-white-fade: var(--border-white-fade) !global;
+  $border-gray-light: var(--border-gray-light) !global;
+  $border-green-light: var(--border-green-light) !global;
+  $border-red-light: var(--border-red-light) !global;
+  $border-yellow: var(--border-yellow) !global;
+  $border-color-button: var(--border-color-button) !global;
+  $bg-black-fade: var(--bg-black-fade) !global;
+  $border-white: var(--border-white) !global;
+  $border-gray: var(--border-gray) !global;
+  $border-gray-dark: var(--border-gray-dark) !global;
+  $border-gray-darker: var(--border-gray-darker) !global;
+  $border-blue: var(--border-blue) !global;
+  $border-blue-light: var(--border-blue-light) !global;
+  $border-green: var(--border-green) !global;
+  $border-purple: var(--border-purple) !global;
+  $border-red: var(--border-red) !global;
+  $bg-white: var(--bg-white) !global;
+  $bg-black: var(--bg-black) !global;
+  $bg-blue-light: var(--bg-blue-light) !global;
+  $bg-blue: var(--bg-blue) !global;
+  $bg-gray-dark: var(--bg-gray-dark) !global;
+  $bg-gray-light: var(--bg-gray-light) !global;
+  $bg-gray: var(--bg-gray) !global;
+  $bg-green: var(--bg-green) !global;
+  $bg-green-light: var(--bg-green-light) !global;
+  $bg-orange: var(--bg-orange) !global;
+  $bg-purple: var(--bg-purple) !global;
+  $bg-purple-light: var(--bg-purple-light) !global;
+  $bg-pink: var(--bg-pink) !global;
+  $bg-red: var(--bg-red) !global;
+  $bg-red-light: var(--bg-red-light) !global;
+  $bg-yellow: var(--bg-yellow) !global;
+  $bg-yellow-light: var(--bg-yellow-light) !global;
+  $bg-yellow-dark: var(--bg-yellow-dark) !global;
+  $text-black: var(--text-black) !global;
+  $text-white: var(--text-white) !global;
+  $text-blue: var(--text-blue) !global;
+  $text-gray-dark: var(--text-gray-dark) !global;
+  $text-gray-light: var(--text-gray-light) !global;
+  $text-gray: var(--text-gray) !global;
+  $text-green: var(--text-green) !global;
+  $text-orange: var(--text-orange) !global;
+  $text-orange-light: var(--text-orange-light) !global;
+  $text-purple: var(--text-purple) !global;
+  $text-pink: var(--text-pink) !global;
+  $text-red: var(--text-red) !global;
+  $text-yellow: var(--text-yellow) !global;
+}
